pyproject.toml
[tool.poetry]
name = "platform-electronics"
version = "0.1.0"
description = ""
authors = ["donkey4hot <tractorist666@gmail.com>"]
readme = "README.md"
packages = [{include = "platform_electronics"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
